Node/services/speech_to_text/providers/nemotron_streaming/requirements.txt
nemo_toolkit[asr]
numpy
